--- src/main/java/Command.java
@@ -1,8 +1,0 @@
-class Command {
-    void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
-    }
-
-    boolean isExit() {
-        return false;
-    }
-}
