--- src/main/java/ExitCommand.java
@@ -1,12 +1,0 @@
-public class ExitCommand extends Command {
-
-    @Override
-    void execute(TaskList tasks, Ui ui, Storage storage) {
-        Ui.showExit();
-    }
-
-    @Override
-    boolean isExit() {
-        return true;
-    }
-}
