+++ src/main/java/duke/Duke.java
@@ -1,3 +1,8 @@
+package duke;
+
+import duke.command.Command;
+import duke.task.TaskList;
+
 public class Duke {
     private Storage storage;
     private TaskList tasks;
