+++ src/main/java/duke/DukeException.java
@@ -1,4 +1,6 @@
-class DukeException extends Exception {
+package duke;
+
+public class DukeException extends Exception {
     DukeException(String message) {
         super(message);
     }
