+++ src/main/java/duke/command/AddDeadlineCommand.java
@@ -1,9 +1,15 @@
+package duke.command;
+
+import duke.*;
+import duke.task.Deadline;
+import duke.task.TaskList;
+
 import java.util.Date;
 
 public class AddDeadlineCommand extends Command {
-    private String description;
-    private String deadlineTime;
-    private Date dueDate;
+    public String description;
+    public String deadlineTime;
+    public Date dueDate;
 
     public AddDeadlineCommand(String description, String deadlineTime) {
         this.description = description;
@@ -16,7 +22,7 @@ public class AddDeadlineCommand extends Command {
     }
 
     @Override
-    void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         Deadline deadline;
         if (dueDate == null) {
             deadline = new Deadline(description, deadlineTime);
