+++ src/main/java/duke/command/MarkCommand.java
@@ -1,3 +1,8 @@
+package duke.command;
+
+import duke.*;
+import duke.task.TaskList;
+
 public class MarkCommand extends Command {
     private int taskNum;
 
@@ -6,7 +11,7 @@ public class MarkCommand extends Command {
     }
 
     @Override
-    void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         if (taskNum - 1 < 0 || taskNum - 1>= tasks.size()) {
             throw new DukeException("\u2639 OOPS!!! The index to mark as done cannot be less than 0 or "
                     + "greater than the length of the list.");
