+++ src/main/java/duke/task/Deadline.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
 import java.util.Date;
