+++ src/main/java/duke/task/TaskList.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 import java.util.ArrayList;
 public class TaskList {
     ArrayList<Task> tasks;
