+++ text-ui-test/EXPECTED.TXT
@@ -1,2 +1,2 @@
-Welcome! I'm Duke.
+Welcome! I'm duke.Duke.
 What can I do for you?
